requirements.txt
streamlit
Pillow
PyMuPDF
pydub
pdf2image
python-docx
ffmpeg-python
